Move evaluate funtion to a helper file so it can be reused by other hooks
