+++ src/utils/hooks/useGetResponsesAndElementState.tsx
@@ -12,9 +12,8 @@ import {
   ResponsesFullByCode,
   TemplateElementState,
   ApplicationElementStates,
-  ElementState,
 } from '../types'
-import evaluateExpression from '@openmsupply/expression-evaluator'
+import evaluateElementExpressions from '../helpers/evaluateElementExpressions'
 
 interface useGetResponsesAndElementStateProps {
   serialNumber: string
@@ -81,50 +80,16 @@ const useGetResponsesAndElementState = (props: useGetResponsesAndElementStatePro
   }, [data, apolloError])
 
   useEffect(() => {
-    if (responsesByCode && Object.keys(responsesByCode).length > 0)
-      evaluateElementExpressions().then((result) => {
+    if (responsesByCode && responsesFullByCode && Object.keys(responsesByCode).length > 0)
+      evaluateElementExpressions({
+        elementsExpressions,
+        responsesByCode,
+        responsesFullByCode,
+      }).then((result) => {
         setElementsState(result)
         setLoading(false)
       })
   }, [responsesByCode])
-
-  async function evaluateElementExpressions() {
-    const promiseArray: Promise<ElementState>[] = []
-    elementsExpressions.forEach((element) => {
-      promiseArray.push(evaluateSingleElement(element))
-    })
-    const evaluatedElements = await Promise.all(promiseArray)
-    const elementsState: ApplicationElementStates = {}
-    evaluatedElements.forEach((element) => {
-      elementsState[element.code] = element
-    })
-    return elementsState
-  }
-
-  async function evaluateSingleElement(element: TemplateElementState): Promise<ElementState> {
-    const evaluationParameters = {
-      objects: [responsesByCode as ResponsesByCode, responsesFullByCode as ResponsesFullByCode], // TO-DO: Also send user/org objects etc.
-      // graphQLConnection: TO-DO
-    }
-    const isEditable = evaluateExpression(element.isEditable, evaluationParameters)
-    const isRequired = evaluateExpression(element.isRequired, evaluationParameters)
-    const isVisible = evaluateExpression(element.visibilityCondition, evaluationParameters)
-    // TO-DO: Evaluate element paremeters (in 'parameters' field, but unique to each element type)
-    const results = await Promise.all([isEditable, isRequired, isVisible])
-    const evaluatedElement = {
-      id: element.id,
-      code: element.code,
-      title: element.title,
-      category: element.category,
-      parameters: element.parameters,
-      elementTypePluginCode: element.elementTypePluginCode,
-      section: element.section as number,
-      isEditable: results[0] as boolean,
-      isRequired: results[1] as boolean,
-      isVisible: results[2] as boolean,
-    }
-    return evaluatedElement
-  }
 
   return {
     error,
